+++ server/curation/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Provide specific routing methods for API endpoints."""
